pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Projeto Bookstore API com Django Rest Framework"
authors = [
    {name = "Tomás Fava",email = "tomsfava@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
pytest-django = "^4.11.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
